+++ TGUI/include/TGUI/EditBox.hpp
@@ -1,7 +1,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //
 // TGUI - Texus's Graphical User Interface
-// Copyright (C) 2012-2013 Bruno Van de Velde (vdv_b@tgui.eu)
+// Copyright (C) 2012-2014 Bruno Van de Velde (vdv_b@tgui.eu)
 //
 // This software is provided 'as-is', without any express or implied warranty.
 // In no event will the authors be held liable for any damages arising from the use of this software.
@@ -27,8 +27,6 @@
 
 
 #include <TGUI/ClickableWidget.hpp>
-
-/// \todo  Double clicking should only select one word in EditBox. To select the whole text, you should click 3 times.
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -391,7 +389,7 @@ namespace tgui
         /// Normally you will not need this function.
         ///
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-        void setSelectionPointPosition(size_t charactersBeforeSelectionPoint);
+        void setSelectionPointPosition(unsigned int charactersBeforeSelectionPoint);
 
 
         /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
